+++ bot/migrations/2019-05-13-225023_index_songs/up.sql
@@ -1,0 +1,1 @@
+CREATE INDEX idx_songs_added_at_id ON songs(added_at, id);